requirements.txt
matplotlib==3.7.0
numpy==1.24.2
scipy==1.10.1
pandas==2.0.0
openpyxl==3.1.2

Pillow~=10.0.0
psutil~=5.9.5
moviepy~=1.0.3
ax~=0.52.0
tqdm~=4.66.1